rename folder SearchPage into search_page_components and search-page into search_page
