+++ client/src/components/search_page_components/Filter.css
@@ -63,7 +63,7 @@
   align-items: center;
   justify-content: center;
   width: 150px;
-  height: 40px;
+  height: 2.1em;
   font-size: 1.5em;
   text-decoration: none;
   padding: 10px 20px;
