+++ client/src/components/search_page_components/HomeStructureList.css
