+++ client/src/main.jsx
@@ -1,19 +1,16 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-
 import {
   createBrowserRouter,
   RouterProvider,
   redirect,
 } from "react-router-dom";
-
 import App from "./App";
 import HomePage from "./pages/home_page/HomePage";
-
 import StructureForm from "./pages/structure_form/StructureForm";
 import ConnexionPage from "./pages/Connexion_page/ConnexionPage";
 import SingIn from "./pages/signin/SignIn";
-import SearchPage from "./pages/search-page/SearchPage";
+import SearchPage from "./pages/search_page/SearchPage";
 import ProfilePage from "./pages/profile_page/ProfilePage";
 
 const URL = import.meta.env.VITE_API_URL;
