+++ client/src/pages/search_page/SearchPage.css
